pyproject.toml
[build-system]
requires = ["hatchling>=1.5.0", "jupyterlab>=4.0.0,<5", "hatch-nodejs-version>=0.3.2"]
build-backend = "hatchling.build"

[project]
name = "dsc10_tutor_jlab_backend"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
classifiers = [
    "Framework :: Jupyter",
    "Framework :: Jupyter :: JupyterLab",
    "Framework :: Jupyter :: JupyterLab :: 4",
    "Framework :: Jupyter :: JupyterLab :: Extensions",
    "Framework :: Jupyter :: JupyterLab :: Extensions :: Prebuilt",
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "google-adk>=1.24.1",
    "google-generativeai>=0.8.6",
    "jupyter_server>=2.4.0,<3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
dynamic = ["version", "description", "authors", "urls", "keywords"]

[dependency-groups]
dev = [
    "babypandas>=1.0.0",
    "jupyterlab~=4.4.0",
    "otter-grader>=6.1.6",
    "pandas~=2.3.3",
    "ruff~=0.13.3",
]

[project.optional-dependencies]
test = [
    "coverage~=7.10.7",
    "pytest~=8.4.2",
    "pytest-asyncio~=1.2.0",
    "pytest-cov~=7.0.0",
    "pytest-jupyter[server]~=0.10.0"
]

[tool.hatch.version]
source = "nodejs"

[tool.hatch.metadata.hooks.nodejs]
fields = ["description", "authors", "urls", "keywords"]

[tool.hatch.build.targets.sdist]
artifacts = ["dsc10_tutor_jlab_backend/labextension"]
exclude = [".github", "binder"]

[tool.hatch.build.targets.wheel.shared-data]
"dsc10_tutor_jlab_backend/labextension" = "share/jupyter/labextensions/dsc10-tutor-jlab-frontend"
"install.json" = "share/jupyter/labextensions/dsc10-tutor-jlab-frontend/install.json"
"jupyter-config/server-config" = "etc/jupyter/jupyter_server_config.d"

[tool.hatch.build.hooks.version]
path = "dsc10_tutor_jlab_backend/_version.py"

[tool.hatch.build.hooks.jupyter-builder]
dependencies = ["hatch-jupyter-builder>=0.5"]
build-function = "hatch_jupyter_builder.npm_builder"
ensured-targets = [
    "dsc10_tutor_jlab_backend/labextension/static/style.js",
    "dsc10_tutor_jlab_backend/labextension/package.json",
]
skip-if-exists = ["dsc10_tutor_jlab_backend/labextension/static/style.js"]

[tool.hatch.build.hooks.jupyter-builder.build-kwargs]
build_cmd = "build:prod"
npm = ["jlpm"]

[tool.hatch.build.hooks.jupyter-builder.editable-build-kwargs]
build_cmd = "install:extension"
npm = ["jlpm"]
source_dir = "src"
build_dir = "dsc10_tutor_jlab_backend/labextension"

[tool.jupyter-releaser.options]
version_cmd = "hatch version"

[tool.jupyter-releaser.hooks]
before-build-npm = [
    "python -m pip install 'jupyterlab>=4.0.0,<5'",
    "jlpm",
    "jlpm build:prod"
]
before-build-python = ["jlpm clean:all"]

[tool.check-wheel-contents]
ignore = ["W002"]
